+++ package.json
@@ -9,12 +9,8 @@
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives",
     "format:fix": "prettier --write src/**/*.{ts,tsx} --fix",
     "preview": "vite preview",
-    "prepare": "husky install"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+    "prepare": "husky install",
+    "lint-staged": "lint-staged"
   },
   "lint-staged": {
     "*.{tsx}": [
